Initial git version of Csmith.
This initial commit corresponds to the Csmith 2.0.0 release.

The git version of Csmith was created by a script that exported the
Csmith project from a Subversion repository at the University of Utah,
from r2202 forward.  Certain files in the Subversion repository ---
files not included in the Csmith 2.0.0 release --- were excluded from
the migration.