Put the current git commit's short hash in generated programs.
We previously put the SVN revision in generated programs.  Now that
the sources are in git, insert the current commit's short hash instead.

Note that I changed `configure.ac' so that we won't get warnings
about using `:=' assignments in Automake files.  I'm not crazy about
this change; better ways to get the current hash at build time would
be appreciated.