Embed Git commit hash by generating "git_version.cpp".
Previously we passed the current git hash to the compiler on the
command line.  In principle, this makes it difficult to know what
needs to be recompiled when the commit hash changes.

Now, generate a file containing the current hash.  When the commit
hash changes, we can just regenerate that file, recompile it, and
relink.