On Windows, add `lrand48` sources to Csmith (Automake) compile.
When using the Automake-based build system to compile Csmith on Windows,
compile and link with an implementation of `lrand48`.

This commit is the Automake equivalent of the CMake changes that I made
earlier (in commit ef1f45a).

Having implemented this for the Automake- and CMake-based build systems,
I'm thinking that I should change the test.  We don't care about the
platform per se; we care whether or not it has `lrand48`.  We should
probably just check for that directly.