Avoid error when using CMake pre-3.0.
Check for policy CMP0042.  If it is not known, don't try to set it.
This check-before-set idiom is suggested by the CMake documentation.
See <https://cmake.org/cmake/help/v3.0/manual/cmake-policies.7.html>.

Without this check, CMake pre-3.0 gives an error.

BTW, did you know that CMake 2.8 is the default on CentOS 7.3?