Remove the least useful code
Eager reducer and simple delta reducer: we have proven them to be less
efficient than CReduce.
Msp filters: I suggest we use a new mechanism for silicon vendors to
restrict which arithmetic operations are supported on their processors.
The mechanism should be through a config file, instead of we creating
vendor specific filters in Csmith.
Splat extension: Splat seems to be less popular than klee and crest
nowadays. Even for klee and crest, we should move them out of Csmith in
the future when we have a generic way to support outputing the random
programs in customized ways.